Hypothesis integration for test_float_img_to_uint8_img. Bug fixed for ivy.float_img_to_unit8_img.

+++ ivy/functional/ivy/image.py
@@ -126,7 +126,7 @@ def float_img_to_uint8_img(x):
         The new encoded uint8 image *[batch_shape,h,w,4]* .
 
     """
-    x_np = ivy.to_numpy(x)
+    x_np = ivy.to_numpy(x).astype('float32')
     x_shape = x_np.shape
     x_bytes = x_np.tobytes()
     x_uint8 = _np.frombuffer(x_bytes, _np.uint8)
